consolidates routes and index.js into app.js -- allows fields property to be passed in upon registration
