+++ server/env.example
@@ -5,5 +5,6 @@ AIRTABLE_ENDPOINT_URL=https://api.airtable.com/
 AIRTABLE_API_VERSION=0.1.0
 AIRTABLE_USER_AGENT=Airtable.js/0.7.1
 AIRTABLE_USER_TABLE=
+AIRTABLE_USERNAME_COLUMN_NAME=
 DISABLE_HASH_PASSWORD=false
 SALT_ROUNDS=
