+++ server/package.json
@@ -2,10 +2,10 @@
   "name": "server",
   "version": "1.0.0",
   "description": "airlock server",
-  "main": "index.js",
+  "main": "app.js",
   "scripts": {
-    "dev": "NODE_ENV=development nodemon index.js",
-    "start": "NODE_ENV=production node index.js"
+    "dev": "NODE_ENV=development nodemon app.js",
+    "start": "NODE_ENV=production node app.js"
   },
   "author": "",
   "license": "ISC",
